Factor out shrinking tests
